+++ src/out_server_message.cpp
@@ -14,8 +14,8 @@ OutServerMessage::~OutServerMessage()
 void OutServerMessage::Send(std::shared_ptr<MsgManager> const& msg_manager)
 {
 	msg_manager->WriteOpcode(GetId());
-	auto len = (uint8_t)(msg.length());
+	auto len = (uint16_t)(msg.length());
 	msg_manager->WriteByte((uint8_t)server_message_type);
-	msg_manager->WriteByte(len);
+	msg_manager->WriteShort(len);
 	msg_manager->WriteString(msg);
 }
